refactor: import advancedQuery URL params from constants file
