+++ website/src/components/genspectrum/advancedQueryUrlParamConstants.ts
@@ -1,0 +1,2 @@
+export const advancedQueryUrlParamForVariant = 'advancedQueryVariant';
+export const advancedQueryUrlParam = 'advancedQuery';
