+++ website/src/views/pageStateHandlers/CompareSideBySidePageStateHandler.ts
@@ -16,7 +16,7 @@ import { parseLocationFiltersFromUrl, setSearchFromLocationFilters } from './loc
 import { decodeFiltersFromSearch, searchParamsFromFilterMap } from './multipleFiltersFromToUrl.ts';
 import { parseNumberRangeFilterFromUrl, setSearchFromNumberRangeFilters } from './numberRangeFilterFromToUrl.ts';
 import { parseTextFiltersFromUrl, setSearchFromTextFilters } from './textFilterFromToUrl.ts';
-import { advancedQueryUrlParam } from '../../components/genspectrum/AdvancedQueryFilter.tsx';
+import { advancedQueryUrlParam } from '../../components/genspectrum/advancedQueryUrlParamConstants.ts';
 import { formatUrl } from '../../util/formatUrl.ts';
 
 export class CompareSideBySideStateHandler implements PageStateHandler<CompareSideBySideData> {
