+++ website/src/views/pageStateHandlers/SingleVariantPageStateHandler.ts
@@ -16,7 +16,7 @@ import { parseLocationFiltersFromUrl, setSearchFromLocationFilters } from './loc
 import { parseNumberRangeFilterFromUrl, setSearchFromNumberRangeFilters } from './numberRangeFilterFromToUrl.ts';
 import { parseTextFiltersFromUrl, setSearchFromTextFilters } from './textFilterFromToUrl.ts';
 import { toLapisFilterWithoutVariant } from './toLapisFilterWithoutVariant.ts';
-import { advancedQueryUrlParam } from '../../components/genspectrum/AdvancedQueryFilter.tsx';
+import { advancedQueryUrlParam } from '../../components/genspectrum/advancedQueryUrlParamConstants.ts';
 import { formatUrl } from '../../util/formatUrl.ts';
 
 export class SingleVariantPageStateHandler<
